Painting Mechanic Mostly Implemented
Players now have the ability to draw on the canvas. However there is currently a bug.

I believe it has to do with the UVs of the model not being that good. There is a tiling issue with the canvas so players cannot freely draw on the canvas and are instead restricted to a smaller area. Removing the "fancy" sections of the material so only the canvas remains still results in the tiling issue, this is what leads me to believe it is the UVs that are the issue and it is not an issue with the material/texture.

Alternatively this could be an issue with the UV channel that the script uses to execute the draw function. The blueprint is currently using channel 1 for the UVs since they seemed decent enough in engine, however this channel could be what is potentially causing the tiling issue