Modular Paint Mechanic Work + Updates
For whatever reason the paint mechanic isn't transferring over to the meshes of the painting. I need to re-evaluate my code and see if something is going wrong there. The sphere highlight is displaying and it registers when each mouse button is held down, however the mesh is not being drawn  on

Aside from that I added a few things for the project in general. I created a game manager with the frame work to take different values of scores. I also added a blueprint which will randomize the paintings if they aren't defined.