This is a question for the professor
Currently the mechanic works part of the way. Without completely redoing the mechanic I do have some semblance of it working correctly. There is a 3D object with a translucent material, that material has a render canvas which allows the user to paint on it. When the user paints on the painting it doesn't paint; however since all canvases are linked when other surfaces are painted and the game returns to the editor only then are marking visible. (Markings are not added during the game however.)