Set Up Very Basic Multiplayer Framework
Player 2 still needs to directly be able to influence what the player has already painted. There is also a slight bug where projectiles will either duplicate themselves or point themselves directly  in front of the player instead of where the player was actually looking.

The issue with the removing decals is the way I was able to remove them in the first place. On event hit, the blueprints are able to grab a scene component to themselves and using this information I am able to grab children of the BPs without affecting anything else however when another object tries to access this information the game registers all children of the object including the actual blueprint itself